% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inprod_mat}
\alias{inprod_mat}
\title{Calculate the inner products of a predictor and coefficient matrix}
\usage{
inprod_mat(X, B)
}
\arguments{
\item{X}{numeric matrix. Predictors.}

\item{B}{numeric matrix. Coefficients.}
}
\description{
Given a matrix of predictors, X, and matrix of coefficients, B, calculate
the inner products, X[i, ] %*%  B[j, ].
}

