% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msm.r
\name{msm}
\alias{msm}
\title{Fit a multispecies model}
\usage{
msm(y, sites, x, species, traits, data, site_re = FALSE, method = c("glmer",
  "jags"), ...)
}
\arguments{
\item{y}{Character. Column name of response variable, species occurrences.}

\item{sites}{Character. Column name of site variable.}

\item{x}{Character. Column names of site level predictor variables.}

\item{species}{Character. Column name of species variable.}

\item{traits}{Character. Column names of species trait variables.}

\item{data}{A data.frame containing the variables for the model.}

\item{site_re}{Logical. Should a site level random effect be included.}

\item{method}{The used to fit the model.}

\item{...}{Further arguments to pass on to model fitting functions.}
}
\description{
Fit a multispecies model
}
\examples{
msm_glmer <- msm('present', 'plot', 'logit_rock', 'species', 'ln_sla', eucs)
msm_glmer_probit <- msm('present', 'plot', 'logit_rock', 'species', 'ln_sla',
  eucs, family=binomial(link='probit'))
msm_jags <- msm('present', 'plot', 'logit_rock', 'species', 'ln_sla', eucs,
  method='jags')
}

