% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/te_plot.r
\name{te_plot}
\alias{te_plot}
\title{Plot the relationship between a trait and the environment infered from a multispecies model}
\usage{
te_plot(x, x_var, trait, nsims = 200)
}
\arguments{
\item{x}{An object of class 'glmerMod'}

\item{x_var}{Character. Name of site-level predictor variable.}

\item{trait}{Character. Name of species trait variable.}

\item{nsims}{Integer. Number of simulations for approximate Bayesian estimates of credible intervals and posterior densities.}
}
\description{
Plot the relationship between a trait and the environment infered from a multispecies model
}
\examples{
library(msm)
data(eucs)
msm_fit <- msm('present', 'plot', 'logit_rock', 'species', 'ln_sla', eucs)
te_plot(msm_fit, 'logit_rock', 'ln_sla')
}

