% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.r
\docType{methods}
\name{waic}
\alias{waic}
\alias{waic,rjags-method}
\alias{waic,rjags.parallel-method}
\title{Calculate waic}
\usage{
waic(x)

\S4method{waic}{rjags}(x)

\S4method{waic}{rjags.parallel}(x)
}
\arguments{
\item{x}{An object of class "rjags", or "rjags.parallel"}
}
\description{
Calculate waic
}
\section{Methods (by class)}{
\itemize{
\item \code{rjags}: calculate waic for rjags and rjags.parallel objects

\item \code{rjags.parallel}: calculate waic for rjags and rjags.parallel objects
}}
\examples{
msm_jags <- msm('present', 'plot', 'logit_rock', 'species', data = eucs, type = 'jsdm',
  method = 'jags')
waic(msm_jags)
}

