% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_plot.r
\docType{methods}
\name{coef_plot}
\alias{coef_plot}
\alias{coef_plot,glmerMod-method}
\title{Plot the coefficients of a multispecies model}
\usage{
coef_plot(x)

\S4method{coef_plot}{glmerMod}(x)
}
\arguments{
\item{x}{An object of class 'glmerMod'}
}
\description{
Plot the coefficients of a multispecies model
}
\section{Methods (by class)}{
\itemize{
\item \code{glmerMod}: coefficient plot for glmer model
}}
\examples{
msm_fit <- msm('present', 'plot', 'logit_rock', 'species', 'ln_sla', eucs)
coef_plot(msm_fit)
}

