% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coef_plot.r
\name{coef_plot}
\alias{coef_plot}
\title{Plot the coefficients of a multispecies model}
\usage{
coef_plot(x)
}
\arguments{
\item{x}{An object of class 'glmerMod'}
}
\description{
Plot the coefficients of a multispecies model
}
\examples{
library(msm)
data(eucs)
msm_fit <- msm('present', 'plot', 'logit_rock', 'species', 'ln_sla', eucs)
coef_plot(msm_fit)
}

