% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msm.r
\name{msm}
\alias{msm}
\title{Fit a multispecies model}
\usage{
msm(y, sites, x, species, traits, data, site_re = FALSE)
}
\arguments{
\item{y}{Character. Column name of response variable, species occurrences.}

\item{sites}{Character. Column name of site variable.}

\item{x}{Character. Column names of site level predictor variables.}

\item{species}{Character. Column name of species variable.}

\item{traits}{Character. Column names of species trait variables.}

\item{data}{A data.frame containing the variables for the model.}

\item{site_re}{Logical. Should a site level random effect be included.}
}
\description{
Fit a multispecies model
}
\examples{
library(msm)
data(eucs)
msm_fit <- msm('present', 'plot', 'logit_rock', 'species', 'ln_sla', eucs)
}

