% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.r
\docType{methods}
\name{waic}
\alias{waic}
\alias{waic,jagsUI-method}
\title{Calculate waic}
\usage{
waic(x)

\S4method{waic}{jagsUI}(x)
}
\arguments{
\item{x}{An object of class "jagsUI"}
}
\description{
Calculate waic
}
\section{Methods (by class)}{
\itemize{
\item \code{jagsUI}: waic for jags model
}}
\examples{
msm_jags <- msm('present', 'plot', 'logit_rock', 'species', data = eucs, type = 'jsdm',
  method = 'jags')
waic(msm_jags)
}

